requirements.txt
amqp==5.3.1
aniso8601==9.0.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
beatbox==32.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==44.0.0
daphne==4.1.2
Deprecated==1.2.15
Django==5.1.3
django-appconf==1.0.6
django-celery-beat==2.7.0
django-celery-email==3.0.0
django-celery-results==2.5.1
django-cors-headers==4.6.0
django-crontab==0.7.1
django-redis==5.4.0
django-timezone-field==7.0
Django-Verify-Email==3.0.3
django-windowsauth==1.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
flower==2.0.1
graphene==3.4.3
graphene-django==3.2.2
graphql-core==3.2.5
graphql-relay==3.2.0
http_ece==1.2.1
humanize==4.11.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
ldap3==2.9.1
msgpack==1.1.0
packaging==24.2
pillow==11.0.0
prometheus_client==0.21.0
promise==2.3
prompt_toolkit==3.0.48
psycopg2==2.9.10
py-vapid==1.9.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==24.3.0
pyparsing==3.2.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.21.0
service-identity==24.2.0
six==1.16.0
sqlparse==0.5.2
stripe==11.3.0
text-unidecode==1.3
tornado==6.4.2
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
wfastcgi==3.0.0
wrapt==1.17.0
zope.interface==7.2
